Rename geppetto-js to src #252
