+++ geppetto-core/src/Utility.js
